Support interleaving hardware threads and CPU cores
It's more interesting to show hardware threads grouped before adjacent
cores because their latency is smaller, and it makes it much more
obvious how the cores communicate. Most often the second thread of
each core appears after the whole list of first threads for all cores,
so that disabling them in the BIOS doesn't change the numbering. So here
when "-t" is passed to request grouping by threads, the numbering is
interleaved (e.g. 0 4 1 5 2 6 3 7).