Add indications in the gnuplot output to reduce font size
Displaying heat maps for 48 CPUs shows an unreadable map made of too
large numbers for the cells. Of course the appearance will also depend
on the terminal, but it's trivial to provide hints to the user by just
emitting commented out statements to change the ticks font size and the
labels font size. Showing 48 cores with 5-pt labels and 7-pt ticks is
perfectly readable in less than 1000x600 output.