Also implement write-exclusive access measurements using CAS
Compare-and-swap (CAS) operations are common with locks and extremely
representative of highly-contented operations that modern CPUs are
facing. Some CPUs even try to enforce extra fairness on such operations.
Here when -w is passed, a single atomic is incremented in turn by each
thread, which exactly corresponds to the attempt to take a lock. The
second counter (seq2) is only used to indicate readiness so that we
don't start the timer before the other thread is ready.

It's interesting to note that the results are often better than with
the default test, because a single cache line is touched, while the
default test requires to sequentially request access to two distinct
cache lines, one in shared mode and the other one in exclusive mode.