Add a preheating option to allow the CPU to switch frequency
Running on CPUs with cpufreq enabled usually shows slower exchanges for
cores that have been idle for some time (usually the first ones but not
only). Here we add an option "-H" to enable pre-heating of the cores
after pinning the thread to them. It's just a loop that waits until it
reaches 200 ms, which is usually sufficient to bring cores up to full
speed. It can add significant measurement time with large core counts,
so only use it when needed.